refactor: split ui site manager
